# docs/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "智能站内邮件系统接口文档最终确认版。当前系统为站内邮件系统，不接入 SMTP/POP3/IMAP。所有接口统一以 /api 为前缀，受保护接口使用 Bearer Token。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/users/me": {
      "get": {
        "summary": "获取当前登录用户信息 P0",
        "deprecated": false,
        "description": "获取当前登录用户信息接口，属于 P0。\n\n业务规则：\n- 必须携带有效 Bearer Token。\n- 返回当前登录用户展示所需信息。\n- 当前最终版要求 `username`、`nickname`、`emailAddress`、`avatarText` 全部返回。\n- 接口层不暴露 `userId`。",
        "operationId": "getCurrentUser",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                },
                "examples": {
                  "success": {
                    "summary": "获取当前用户成功",
                    "value": {
                      "code": 0,
                      "message": "success",
                      "data": {
                        "username": "admin",
                        "nickname": "管理员",
                        "emailAddress": "admin@mail.com",
                        "avatarText": "管"
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notLogin": {
                    "summary": "未携带 token",
                    "value": {
                      "code": 40002,
                      "message": "请先登录",
                      "data": null
                    }
                  },
                  "invalidToken": {
                    "summary": "token 无效或过期",
                    "value": {
                      "code": 40002,
                      "message": "未登录或 token 无效",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "系统内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "serverError": {
                    "summary": "系统内部错误",
                    "value": {
                      "code": 50000,
                      "message": "系统内部错误",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/settings": {
      "get": {
        "summary": "获取用户设置和 AI 配置状态 P1",
        "deprecated": false,
        "description": "获取用户设置和 AI 配置状态接口，属于 P1。\n\n业务规则：\n- 必须携带有效 Bearer Token。\n- 用于设置页回显 AI 分析开关、自动回复建议开关、优先级排序开关和模型配置状态。\n- 后端不得返回完整 API Key，只能返回 `apiKeyConfigured` 和脱敏后的 `maskedApiKey`。\n- 未配置模型时，`provider`、`modelName`、`baseUrl`、`maskedApiKey` 可为 null，但字段仍应返回。",
        "operationId": "getUserSettings",
        "tags": [
          "用户设置"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notLogin": {
                    "summary": "未携带 token",
                    "value": {
                      "code": 40002,
                      "message": "请先登录",
                      "data": null
                    }
                  },
                  "invalidToken": {
                    "summary": "token 无效或过期",
                    "value": {
                      "code": 40002,
                      "message": "未登录或 token 无效",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "系统内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "serverError": {
                    "summary": "系统内部错误",
                    "value": {
                      "code": 50000,
                      "message": "系统内部错误",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新用户设置和 AI 配置 P1",
        "deprecated": false,
        "description": "更新用户设置和 AI 模型配置接口，属于 P1。\n\n业务规则：\n- 必须携带有效 Bearer Token。\n- 请求体采用一个统一对象，所有字段在 OpenAPI 层面均非必填。\n- 只更新开关时，可只提交 `aiEnabled`、`autoReplyEnabled`、`prioritySortEnabled`。\n- 配置模型时，`provider`、`baseUrl`、`modelName` 条件必填。\n- 首次配置模型时 `apiKey` 条件必填。\n- `apiKey` 不传表示不修改旧 Key；传空字符串 `\"\"` 表示清空已保存 Key。\n- 当前版本只适配 OpenAI Compatible 模型服务。",
        "operationId": "updateUserSettings",
        "tags": [
          "用户设置"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unsupportedProvider": {
                    "summary": "provider 不支持",
                    "value": {
                      "code": 40000,
                      "message": "不支持的 AI 服务商",
                      "data": null
                    }
                  },
                  "baseUrlEmpty": {
                    "summary": "Base URL 为空",
                    "value": {
                      "code": 40000,
                      "message": "Base URL 不能为空",
                      "data": null
                    }
                  },
                  "baseUrlInvalid": {
                    "summary": "Base URL 格式不合法",
                    "value": {
                      "code": 40000,
                      "message": "Base URL 格式不合法",
                      "data": null
                    }
                  },
                  "modelNameEmpty": {
                    "summary": "模型名称为空",
                    "value": {
                      "code": 40000,
                      "message": "模型名称不能为空",
                      "data": null
                    }
                  },
                  "apiKeyEmptyOnFirstConfig": {
                    "summary": "首次配置时 API Key 为空",
                    "value": {
                      "code": 40000,
                      "message": "API Key 不能为空",
                      "data": null
                    }
                  },
                  "temperatureInvalid": {
                    "summary": "temperature 超出范围",
                    "value": {
                      "code": 40000,
                      "message": "temperature 参数不合法",
                      "data": null
                    }
                  },
                  "maxTokensInvalid": {
                    "summary": "maxTokens 超出范围",
                    "value": {
                      "code": 40000,
                      "message": "maxTokens 参数不合法",
                      "data": null
                    }
                  },
                  "timeoutMsInvalid": {
                    "summary": "timeoutMs 超出范围",
                    "value": {
                      "code": 40000,
                      "message": "timeoutMs 参数不合法",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notLogin": {
                    "summary": "未携带 token",
                    "value": {
                      "code": 40002,
                      "message": "请先登录",
                      "data": null
                    }
                  },
                  "invalidToken": {
                    "summary": "token 无效或过期",
                    "value": {
                      "code": 40002,
                      "message": "未登录或 token 无效",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "系统内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "serverError": {
                    "summary": "系统内部错误",
                    "value": {
                      "code": 50000,
                      "message": "系统内部错误",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "业务错误码。0 表示成功，非 0 表示失败。"
          },
          "message": {
            "type": "string",
            "description": "错误提示，可直接用于前端提示或二次映射。"
          },
          "data": {
            "description": "错误时通常为 null。",
            "type": "null"
          }
        },
        "example": {
          "code": 40001,
          "message": "用户名或密码错误",
          "data": null
        }
      },
...
      "CurrentUserVO": {
        "type": "object",
        "required": [
          "username",
          "nickname",
          "emailAddress",
          "avatarText"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "管理员"
          },
          "emailAddress": {
            "type": "string",
            "description": "系统内邮箱地址",
            "example": "admin@mail.com"
          },
          "avatarText": {
            "type": "string",
            "description": "前端头像占位文字",
            "example": "管"
          }
        },
        "additionalProperties": false
      },
...
      "CurrentUserResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/CurrentUserVO"
          }
        }
      },
...
      "Provider": {
        "type": "string",
        "enum": [
          "qwen",
          "openai",
          "deepseek",
          "kimi",
          "glm",
          "siliconflow",
          "custom"
        ]
      },
      "UserSettingsVO": {
        "type": "object",
        "required": [
          "aiEnabled",
          "autoReplyEnabled",
          "prioritySortEnabled",
          "modelConfigured",
          "provider",
          "modelName",
          "baseUrl",
          "apiKeyConfigured",
          "maskedApiKey",
          "timeoutMs",
          "maxTokens",
          "temperature"
        ],
        "properties": {
          "aiEnabled": {
            "type": "boolean",
            "description": "是否启用 AI 分析",
            "example": true
          },
          "autoReplyEnabled": {
            "type": "boolean",
            "description": "是否启用自动回复建议",
            "example": true
          },
          "prioritySortEnabled": {
            "type": "boolean",
            "description": "是否启用邮件优先级排序",
            "example": true
          },
          "modelConfigured": {
            "type": "boolean",
            "description": "为 true 表示 provider、baseUrl、modelName 均已配置且 apiKeyConfigured = true",
            "example": true
          },
          "provider": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Provider"
              },
              {
                "type": "null"
              }
            ],
            "description": "当前模型服务商；未配置模型时为 null"
          },
          "modelName": {
            "type": "string",
            "description": "当前模型名称",
            "nullable": true,
            "example": "deepseek-chat"
          },
          "baseUrl": {
            "type": "string",
            "description": "当前模型接口地址",
            "nullable": true,
            "example": "https://api.deepseek.com"
          },
          "apiKeyConfigured": {
            "type": "boolean",
            "description": "是否已保存 API Key",
            "example": true
          },
          "maskedApiKey": {
            "type": "string",
            "description": "脱敏后的 API Key",
            "nullable": true,
            "example": "sk-****abcd"
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "description": "单次 AI 请求超时时间，单位毫秒",
            "example": 10000
          },
          "maxTokens": {
            "type": "integer",
            "format": "int32",
            "description": "模型最大输出长度",
            "example": 800
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "模型温度",
            "example": 0.2
          }
        },
        "additionalProperties": false
      },
      "UserSettingsResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/UserSettingsVO"
          }
        }
      },
      "UpdateUserSettingsRequest": {
        "type": "object",
        "description": "所有字段均为非必填。配置模型时 provider、baseUrl、modelName 条件必填；首次配置模型时 apiKey 必填。apiKey 不传表示不修改旧 Key；apiKey 为空字符串表示清空已保存 Key。",
        "properties": {
          "aiEnabled": {
            "type": "boolean",
            "description": "是否启用 AI 功能",
            "example": true
          },
          "autoReplyEnabled": {
            "type": "boolean",
            "description": "是否启用回复建议",
            "example": true
          },
          "prioritySortEnabled": {
            "type": "boolean",
            "description": "是否启用优先级排序",
            "example": true
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "apiKey": {
            "type": "string",
            "description": "新的 API Key；不传表示不修改旧 Key；空字符串表示清空 Key",
            "example": "sk-xxxxxxxxxxxxxxxx"
          },
          "baseUrl": {
            "type": "string",
            "description": "OpenAI 兼容接口地址",
            "example": "https://api.deepseek.com"
          },
          "modelName": {
            "type": "string",
            "description": "模型名称",
            "example": "deepseek-chat"
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "description": "超时时间，默认 10000",
            "example": 10000
          },
          "maxTokens": {
            "type": "integer",
            "format": "int32",
            "description": "最大输出长度，默认 800",
            "example": 800
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "温度参数，默认 0.2",
            "example": 0.2
          }
        },
        "additionalProperties": false
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}